+++ site/src/examples/omezarr/render-utils.ts
@@ -28,7 +28,8 @@ type Thing = {
     settings: RenderSettings;
 };
 function mapValues<T extends Record<string, V>, V, R>(obj: T, fn: (v: V) => R): { [k in keyof T]: R } {
-    return (Object.keys(obj) as (keyof T)[]).reduce( // typecast is annoyingly necessary in this case to avoid a linting warning
+    return (Object.keys(obj) as (keyof T)[]).reduce(
+        // typecast is annoyingly necessary in this case to avoid a linting warning
         (acc, k) => {
             acc[k] = fn(obj[k]);
             return acc;
